requirements.txt
tensorflow
keras
numpy
pillow
matplotlib
fastapi
uvicorn
python-multipart
gdown
scikit-learn
nltk
pyttsx3
